User_Auth_connect_with_Thread_id

+++ app/Http/Controllers/ThreadController.php
@@ -7,6 +7,11 @@ use Illuminate\Http\Request;
 
 class ThreadController extends Controller
 {
+    function __construct()
+    {
+        return $this->middleware('auth')->except('index');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -45,8 +50,7 @@ class ThreadController extends Controller
 
         ]);
         //        //Store
-
-        Thread::create($request->all());
+        auth()->user()->threads()-> create($request->all());
 
         //        //Redirect
 
@@ -85,6 +89,11 @@ class ThreadController extends Controller
      */
     public function update(Request $request, Thread $thread)
     {
+        if (auth()->user()->id !== $thread->user_id){
+
+            abort(401,"unauthorized");
+
+        }
         $this-> validate($request,[
 
             'subject'=>'required|min:10',
@@ -92,6 +101,8 @@ class ThreadController extends Controller
             'thread'=>'required|min:20',
 
         ]);
+
+
 
         //update
         $thread->update($request->all());
@@ -107,6 +118,11 @@ class ThreadController extends Controller
      */
     public function destroy(Thread $thread)
     {
+        if (auth()->user()->id !== $thread->user_id){
+
+            abort(401,"unauthorized");
+
+        }
         $thread->delete();
 
         return redirect()->route('thread.index')->withMessage("Thread Deleted");
